+++ src/cppNGSD/GermlineReportGenerator.cpp
@@ -466,7 +466,7 @@ void GermlineReportGenerator::writeHTML(QString filename)
 			QString trait = row[trait_idx];
 			QString score = row[score_idx];
 			QString zscore = "n/a";
-			QString population = processed_sample_data.ancestry;
+			QString population = NGSHelper::populationCodeToHumanReadable(processed_sample_data.ancestry);
 			if (trait=="Breast Cancer") // mean and standard deviation taken from BCAC315 data
 			{
 				double mean = -0.424;
@@ -817,7 +817,7 @@ void GermlineReportGenerator::writeXML(QString filename, QString html_document)
 
 		//element ExternalLink
 		w.writeStartElement("ExternalLink");
-		w.writeAttribute("url", "https://dgv.tcag.ca/gb2/gbrowse/dgv2_"+buildToString(data_.build)+"/?name=" + cnv.toString());
+		w.writeAttribute("url", "http://dgv.tcag.ca/gb2/gbrowse/dgv2_"+buildToString(data_.build)+"/?name=" + cnv.toString());
 		w.writeAttribute("type", "DGV");
 		w.writeEndElement();
 		w.writeStartElement("ExternalLink");