+++ doc/GSvar/changelog.md
@@ -1,8 +1,12 @@
 # GSvar change log
 
+### 2021_06-38 (13.07.21)
+- small variants search dialog: added maximum NGSD count filter.
+- CNV search dialog: added option to scale log-likelihood, added validation column, added report config comments column.
+- SV search dialog: added validation column, added report config comments column.
+
 ### 2021_06-9 (22.06.21)
 - added warning dialog shown after loading a GSvar file if sample quality is 'bad', or KASP swap probability is larger than 3%.
-
 
 ### 2021_03-69 (20.05.21)
 - General: added pre-filtering of CNVs if too many.